server/requirements.txt
openai
elevenlabs
python-dotenv
google-cloud-firestore